Create HomeContext instead of default context
