+++ src/context/HomeContext/index.tsx
@@ -1,0 +1,1 @@
+export { HomeContext, HomeProvider } from './HomeContext';
